Graph Bug Fixes
* add subpixel draw
* draw_text return text height
* update & fix progress control
* separate active & resize window message
* stristr
* fix canvas size for canvas area
* fix overflow font drawing - canvas sz
* change from const char to char in text_parse_next
* update qnx hid driver

+++ include/aroma/graph/text.h
@@ -240,10 +240,10 @@ byte libaroma_text_draw_ex(
 
 /*
  * Function    : libaroma_draw_text_ex
- * Return Value: byte
+ * Return Value: int
  * Descriptions: direct text directly from string - extended
  */
-byte libaroma_draw_text_ex(
+int libaroma_draw_text_ex(
   LIBAROMA_CANVASP dest,
   const char * text,
   int x, int y,
